pyproject.toml
[project]
name = "usb-agents"
version = "0.1.0"
description = "MCP interop bake-off harness for OpenAI, Microsoft, and Mistral agent runtimes."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "jsonschema>=4.21.1",
  "click>=8.1.0",
  "openai>=1.30.0",
  "azure-core>=1.30.0",
  "mistralai>=0.4.0",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0",
  "opentelemetry-instrumentation>=0.46b0",
  "pandas>=2.2.0",
  "matplotlib>=3.8.0",
  "tenacity>=8.3.0",
  "pyyaml>=6.0.0",
  "rich>=13.7.0",
  "tabulate>=0.9.0",
  "numpy>=1.26.0",
  "jinja2>=3.1.0",
  "anyio>=4.3.0",
  "typer>=0.12.3",
  "pyflakes>=3.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
]

[project.urls]
homepage = "https://example.com/usb-agents"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-q --disable-warnings"
pythonpath = [
  "."
]
testpaths = [
  "tests"
]

[tool.mypy]
python_version = "3.10"
strict = false

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
